Spin flywheel while driving in autonomous
This commit refactors the autonomous routines to spin up the flywheel motor concurrently with following the path to the shooting location. This is achieved by wrapping the `spin` and `FollowPath` commands in a `ParallelGroup`.

The delays after path following have also been increased to allow the flywheel more time to reach its target speed before kicking. The delays have been adjusted from `1.5` or `2.0` seconds to `2.5` seconds.